core/requirements.txt
djangorestframework==3.13.1
Django==4.0.6
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0